Note about not using mass assignment
Added a note about not using mass assignment before we get to the reading on Strong Params. A few students were trying to use it since they got used to it in Sinatra. One even looked up strong params before he got to the reading because he assumed that was the way to solve it rather than passing in the attributes individually. Adding this note here should cover the next few labs, but may want to just add it onto each lab to be clear until we get to strong params.